feat: Fix release job (#20)
Admittedly this took me a long time to get right. I was tripping over
having to have a matrix of build jobs, but a single deploy job. In the
end it was pretty obvious, but there was a lot of stumbling along the
way. I've since cleaned up those temporary builds and tags, but some of
it might be lurking around.