feat: Import files (#55)
Was a bit of a pain to get here. I want to use `libarchive`, but getting
that compiled for Windows was a pain (it was doubling / quadrupling the
build time, and still failing. And the macOS build dynamically linked an
Intel Homebrew path, which crashed the app when launching on an arm
mac). So I ended up going with tried and true `7za`, from
https://github.com/develar/7zip-bin

This also tweaks the onboarding alerts. Since people can just drag in
files, and the app automatically reloads when that happens, it's a
little more straightforward.

Closes https://github.com/mikew/wadpunk/issues/25

## Preview



https://github.com/mikew/wadpunk/assets/4729/4b27b4b8-f9e2-4814-afce-ad53c7ef0bfc