feat: WIP (#88)
- add open game folder button to game dialog
- refactor startGame so it only needs a game id (so there can be a start
game button directly in the game list)

## Notes

I didn't intend for it, but this ended up being a "vibes based coding"
feature. I've been experimenting with openhands at work, and decided to
try it out on a refactor I've tried but got too bored with. There's
still some manual stuff in there, but for the most part it was
openhands/claude

## Preview



https://github.com/user-attachments/assets/3a31566c-9127-43d7-b4d9-a9e263a674ae